OpenAI-sdk-chatbot/pyproject.toml
[project]
name = "vs-labs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.49.1",
    "openai-agents>=0.4.1",
    "python-dotenv>=1.1.1",
]
